# backend/docs/swagger-output.yaml
openapi: 3.0.3
info:
  title: Multi-Tenant Gateway API
  version: 1.0.0
  description: API untuk autentikasi, manajemen tenant, service gateway, dan proxy antar layanan.
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    superadminToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    userToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Error message here
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, tenant]
              properties:
                username: { type: string, example: superadmin }
                password: { type: string, example: supersecretpassword }
                tenant: { type: string, example: default }
      responses:
        "200":
          description: Login sukses
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      tags: [Auth]
      summary: Registrasi user baru (oleh superadmin)
      security:
        - superadminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, role]
              properties:
                username: { type: string, example: admin }
                password: { type: string, example: admin }
                role: { type: string, example: admin }
      responses:
        "200": { description: User registered }
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...